SandboxAPITransmodelApi: restore existing functionality
