+++ src/ext/graphql/transmodelapi/schema.graphql
@@ -587,7 +587,6 @@ type RoutingParameters {
     allowBikeRental: Boolean
     "Separate cost for boarding a vehicle with a bicycle, which is more difficult than on foot."
     bikeBoardCost: Int
-    bikeParkAndRide: Boolean
     "Cost to park a bike."
     bikeParkCost: Int
     "Time to park a bike."
