+++ src/ext/java/org/opentripplanner/ext/transmodelapi/model/EnumTypes.java
@@ -83,7 +83,7 @@ public class EnumTypes {
             .name("VertexType")
             .value("normal", VertexType.NORMAL)
             .value("transit", VertexType.TRANSIT)
-            .value("vehiclePark", VertexType.VEHICLEPARK)
+            .value("bikePark", VertexType.VEHICLEPARK)
             .value("bikeShare", VertexType.BIKESHARE)
             //TODO QL: .value("parkAndRide", VertexType.PARKANDRIDE)
             .build();
