+++ src/ext/java/org/opentripplanner/ext/transmodelapi/model/stop/PlaceInterfaceType.java
@@ -55,8 +55,11 @@ public class PlaceInterfaceType {
             return schema.getObjectType("RentalVehicle");
           }
           if (o instanceof VehicleParking) {
-            return schema.getObjectType(VehicleParkingType.NAME);
+            return schema.getObjectType(BikeParkType.NAME);
           }
+          //if (o instanceof CarPark) {
+          //    return (GraphQLObjectType) carParkType;
+          //}
           return null;
         })
         .build();
